+++ src/main/java/csironi/ggp/course/evalfunctions/EvalZero.java
@@ -24,7 +24,7 @@ public class EvalZero extends EvaluationFunction {
 	 * @see csironi.ggp.course.evalfunctions.EvaluationFunction#eval(org.ggp.base.util.statemachine.MachineState, org.ggp.base.util.statemachine.Role)
 	 */
 	@Override
-	public int eval(ExplicitMachineState state, ExplicitRole role) {
+	public double eval(ExplicitMachineState state, ExplicitRole role) {
 		return 0;
 	}
